Implement fake atomics in runtime4
These are necessary for the atomic externals in stdlib to be usable with runtime4.
They are copies of the versions in runtime/, but assume Domain_alone = true.

Once we implement flambda2 support for the atomic primitives, the externals should
be updated to use them, and these functions may be deleted.