flambda2: Recognize affine switches (#4571)
The middle-end (flambda2) recognizes and transforms identity switches,
boolean not switches, and constant switches that get transformed into
table lookups.

Unlike other parts of the compiler that implement these transformations,
affine switches are not recognized. This means that when flambda2
transforms a switch into a lookup table, later passes are prevented from
using an affine switch instead.

This patch implements the recognition of affine switches in flambda2.

Fixes #4474